DPP-11 #DONE tests created
