+++ service2/application/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask, request
+
+app = Flask(__name__)
+
+from application import routes
