+++ service2/application/routes.py
@@ -1,0 +1,8 @@
+from application import app
+from flask import Flask, request, Response
+import random
+
+@app.route("/randomnames")
+def random_names():
+    namesn= str(random.randint(0,9))+str(random.randint(0,9))
+    return Response(namesn, mimetype='text/plain')
