+++ service3/app.py
@@ -1,11 +1,7 @@
-from flask import Flask
+from flask import Flask, Response
 from random import randint
+from application import app
 
-app = Flask(__name__)
-
-@app.route("/winner")
-def winner():
-    return str(randint(0,7))
 
 if __name__=="__main__":
     app.run(host = "0.0.0.0", port = 5002, debug = True)
